Refactor ETL tests into dedicated package
