+++ backend/tests/etl/__init__.py
@@ -1,0 +1,8 @@
+"""ETL-specific backend tests."""
+
+__all__ = [
+    "test_connection",
+    "test_schema",
+    "test_runner",
+    "test_retry",
+]
